fix: deposit withdraw fixes
