+++ apps/web/src/app/(networks)/(evm)/perps/vaults/[address]/_ui/vault-deposit-dialog.tsx
@@ -10,6 +10,7 @@ import {
   PerpsDialogTrigger,
 } from '@sushiswap/ui'
 import { type ReactNode, useCallback, useMemo, useState } from 'react'
+import { useSpotClearinghouseState } from 'src/lib/perps'
 import { useVaultDetails } from 'src/lib/perps/info/use-vault-details'
 import { useVaultTransfer } from 'src/lib/perps/vaults/use-vault-transfer'
 import { Checker } from 'src/lib/wagmi/systems/Checker'
@@ -17,12 +18,15 @@ import { useAccount } from 'src/lib/wallet'
 import { Amount } from 'sushi'
 import { type EvmAddress, EvmChainId, USDC } from 'sushi/evm'
 import { InputWithKeyboard } from '~evm/perps/_ui/_common'
+import { useUserSettingsState } from '~evm/perps/_ui/account-management'
 import { PerpsChecker } from '~evm/perps/_ui/perps-checker'
+import { useUserState } from '~evm/perps/user-provider'
 
 const currency = USDC[EvmChainId.ARBITRUM]
 const chainId = EvmChainId.ARBITRUM
+const MIN_DEPOSIT_AMOUNT = 5 //5.000000 usdc
 
-export const WithdrawDialog = ({
+export const VaultDepositDialog = ({
   trigger,
   isOpen,
   onOpenChange,
@@ -34,25 +38,46 @@ export const WithdrawDialog = ({
   vaultAddress: EvmAddress
 }) => {
   const { updateVaultTransferAsync, isPending } = useVaultTransfer()
-  const {
-    data: vaultDetails,
-    isLoading,
-    error,
-  } = useVaultDetails({
+  const { data: vaultDetails, refetch } = useVaultDetails({
     vaultAddress,
   })
   const [open, setOpen] = useState<boolean>(false)
   const [amount, setAmount] = useState<string>('')
   const _amount = Amount.tryFromHuman(currency, amount)
-
+  const {
+    state: {
+      webData2Query: {
+        data,
+        isLoading: isLoadingWebData2,
+        error: errorWebData2,
+      },
+    },
+  } = useUserState()
+  const {
+    state: { isUnifiedAccountModeEnabled },
+  } = useUserSettingsState()
   const address = useAccount('evm')
+  const {
+    data: spotClearinghouseState,
+    isLoading: isLoadingSpotClearinghouse,
+    error: errorSpotClearinghouse,
+  } = useSpotClearinghouseState({ address })
+  const isLoading = isLoadingWebData2 || isLoadingSpotClearinghouse
+  const error = errorWebData2 || errorSpotClearinghouse
 
-  const withdrawableBalance = useMemo(() => {
-    if (!vaultDetails?.followerState || !address) return '0'
-    return vaultDetails?.followerState?.vaultEquity
-  }, [vaultDetails?.followerState, address])
+  const depositableBalance = useMemo(() => {
+    if (isUnifiedAccountModeEnabled) {
+      const usdc = spotClearinghouseState?.balances?.find(
+        (b) => b.coin === 'USDC',
+      )
+      const total = usdc?.total || '0'
+      const hold = usdc?.hold || '0'
+      return Number(total) - Number(hold)
+    }
+    return data?.clearinghouseState.withdrawable
+  }, [data, isUnifiedAccountModeEnabled, spotClearinghouseState])
 
-  const balance = Amount.tryFromHuman(currency, withdrawableBalance ?? '0')
+  const balance = Amount.tryFromHuman(currency, depositableBalance ?? '0')
 
   const isControlled = isOpen !== undefined
   const resolvedOpen = isControlled ? isOpen : open
@@ -69,48 +94,32 @@ export const WithdrawDialog = ({
   )
 
   const insufficientBalance =
-    address &&
-    withdrawableBalance &&
-    _amount &&
-    (balance?.lt(_amount) || balance?.eq('0'))
+    address && depositableBalance && _amount && balance?.lt(_amount)
 
-  const withdrawUsdc = useCallback(async () => {
+  const depositUsdc = useCallback(async () => {
     if (!address || !_amount) return
 
     await updateVaultTransferAsync({
-      type: 'withdraw',
+      type: 'deposit',
       usdAmount: _amount.amount.toString(),
       vaultAddress,
     })
+    await refetch()
     setAmount('')
     setOpen(false)
-  }, [address, _amount, vaultAddress, updateVaultTransferAsync])
-
-  const { canWithdraw, unlockDate } = useMemo(() => {
-    if (!vaultDetails?.followerState)
-      return { canWithdraw: false, unlockDate: null }
-
-    const unlockTime = vaultDetails?.followerState?.lockupUntil
-    const canWithdraw = Date.now() >= unlockTime
-    const unlockDate = new Date(unlockTime)
-
-    return { canWithdraw, unlockDate }
-  }, [vaultDetails?.followerState])
+  }, [address, _amount, vaultAddress, updateVaultTransferAsync, refetch])
 
   return (
     <PerpsDialog open={resolvedOpen} onOpenChange={handleOpenChange}>
       <PerpsDialogTrigger asChild>
-        {trigger ? (
-          trigger
-        ) : (
-          <Button variant="perps-secondary">Withdraw</Button>
-        )}
+        {trigger ? trigger : <Button variant="perps-secondary">Deposit</Button>}
       </PerpsDialogTrigger>
       <PerpsDialogContent>
         <PerpsDialogHeader>
-          <PerpsDialogTitle>Withdraw</PerpsDialogTitle>
+          <PerpsDialogTitle>Deposit</PerpsDialogTitle>
           <PerpsDialogDescription>
-            Withdraw funds from {vaultDetails?.name}.
+            Deposit funds to {vaultDetails?.name}. The deposit lock-up period is
+            1 day.
           </PerpsDialogDescription>
         </PerpsDialogHeader>
         <PerpsDialogInnerContent>
@@ -139,67 +148,67 @@ export const WithdrawDialog = ({
                 >
                   <Checker.Custom
                     size="default"
-                    showChildren={Boolean(amount)}
-                    buttonText={'Enter Amount'}
+                    showChildren={vaultDetails?.allowDeposits}
+                    buttonText={'This vault does not accept deposits'}
                     onClick={() => {}}
-                    disabled={!amount}
+                    disabled={!vaultDetails?.allowDeposits}
                     variant="perps-tertiary"
                   >
                     <Checker.Custom
                       size="default"
-                      showChildren={!insufficientBalance}
-                      buttonText={'Insufficient Balance'}
+                      showChildren={Boolean(amount)}
+                      buttonText={'Enter Amount'}
                       onClick={() => {}}
-                      disabled={Boolean(insufficientBalance)}
+                      disabled={!amount}
                       variant="perps-tertiary"
                     >
                       <Checker.Custom
                         size="default"
-                        showChildren={canWithdraw}
-                        buttonText={`Lockup period active`}
+                        showChildren={!insufficientBalance}
+                        buttonText={'Insufficient Balance'}
                         onClick={() => {}}
-                        disabled={!canWithdraw}
+                        disabled={Boolean(insufficientBalance)}
                         variant="perps-tertiary"
                       >
-                        <PerpsChecker.EnableTrading
+                        <Checker.Custom
                           size="default"
+                          showChildren={Number(amount) >= MIN_DEPOSIT_AMOUNT}
+                          buttonText={`Minimum Deposit ${MIN_DEPOSIT_AMOUNT} USDC`}
+                          onClick={() => {}}
+                          disabled={Number(amount) < MIN_DEPOSIT_AMOUNT}
                           variant="perps-tertiary"
                         >
-                          <PerpsChecker.BuilderFee
+                          <PerpsChecker.EnableTrading
                             size="default"
                             variant="perps-tertiary"
                           >
-                            <PerpsChecker.HyperReferral
+                            <PerpsChecker.BuilderFee
                               size="default"
                               variant="perps-tertiary"
                             >
-                              <Button
+                              <PerpsChecker.HyperReferral
                                 size="default"
-                                className="w-full"
-                                onClick={withdrawUsdc}
-                                loading={isPending}
                                 variant="perps-tertiary"
                               >
-                                Withdraw
-                              </Button>
-                            </PerpsChecker.HyperReferral>
-                          </PerpsChecker.BuilderFee>
-                        </PerpsChecker.EnableTrading>
+                                <Button
+                                  size="default"
+                                  className="w-full"
+                                  onClick={depositUsdc}
+                                  loading={isPending}
+                                  variant="perps-tertiary"
+                                >
+                                  Deposit
+                                </Button>
+                              </PerpsChecker.HyperReferral>
+                            </PerpsChecker.BuilderFee>
+                          </PerpsChecker.EnableTrading>
+                        </Checker.Custom>
                       </Checker.Custom>
                     </Checker.Custom>
                   </Checker.Custom>
                 </Checker.Network>
               </Checker.Connect>
             </PerpsChecker.Legal>
-            {!canWithdraw && unlockDate ? (
-              <div>
-                <p className="text-xs text-red-400 text-center italic mb-1">
-                  Withdrawals are disabled for a lockup period after each of
-                  your deposits. You can withdraw after{' '}
-                  {unlockDate.toLocaleString()}.
-                </p>
-              </div>
-            ) : null}
           </div>
         </PerpsDialogInnerContent>
       </PerpsDialogContent>
